requirements.txt
discord.py
rich
aiofiles
dotenv
watchdog